MODIFY, ADD, CHANGE: fuzzMethod() inside applyEnhancedFuzzingToBitmapContextWithFloats()
MODIFY, ADD, CHANGE: fuzzMethod() inside appleEnhancedFuzzingToBitmapContextWithFloats()

            switch (fuzzMethod) {
                case 0: // Additive noise
                    for (int i = 0; i < 4; i++) {
                        float noise = (float)(((double)arc4random() / UINT32_MAX) * 2.0 - 1.0); // Noise range [-1, 1]
                        rawData[pixelIndex + i] += noise;
                    }
                    break;
                case 1: // Multiplicative noise (scale)
                    for (int i = 0; i < 4; i++) {
                        float scale = (float)(((double)arc4random() / UINT32_MAX) * 2.0); // Scale range [0, 2]
                        rawData[pixelIndex + i] *= scale;
                    }
  


#include "..." search starts here:
#include <...> search starts here:
 /usr/local/Cellar/llvm/17.0.6_1/lib/clang/17/include
 /Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS17.2.sdk/usr/include
 /Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS17.2.sdk/System/Library/Frameworks (framework directory)
End of search list.
cx.srd.img-fuzz-001/main.m:246:62: warning: implicit conversion from 'unsigned int' to 'float' changes value from 4294967295 to 4294967296 [-Wimplicit-const-int-float-conversion]
  246 |                         float noise = ((float)arc4random() / UINT32_MAX) * 2.0f - 1.0f; // Noise range [-1, 1]
      |                                                            ~ ^~~~~~~~~~
/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS17.2.sdk/usr/include/stdint.h:109:27: note: expanded from macro 'UINT32_MAX'
  109 | #define UINT32_MAX        4294967295U
      |                           ^~~~~~~~~~~
cx.srd.img-fuzz-001/main.m:252:62: warning: implicit conversion from 'unsigned int' to 'float' changes value from 4294967295 to 4294967296 [-Wimplicit-const-int-float-conversion]
  252 |                         float scale = ((float)arc4random() / UINT32_MAX) * 2.0f; // Scale range [0, 2]
      |                                                            ~ ^~~~~~~~~~
/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS17.2.sdk/usr/include/stdint.h:109:27: note: expanded from macro 'UINT32_MAX'
  109 | #define UINT32_MAX        4294967295U
      |                           ^~~~~~~~~~~
2 warnings generated.

+++ xnuimagefuzzer.m
@@ -243,13 +243,13 @@ void applyEnhancedFuzzingToBitmapContextWithFloats(float *rawData, size_t width,
             switch (fuzzMethod) {
                 case 0: // Additive noise
                     for (int i = 0; i < 4; i++) {
-                        float noise = ((float)arc4random() / UINT32_MAX) * 2.0f - 1.0f; // Noise range [-1, 1]
+                        float noise = (float)(((double)arc4random() / UINT32_MAX) * 2.0 - 1.0); // Noise range [-1, 1]
                         rawData[pixelIndex + i] += noise;
                     }
                     break;
                 case 1: // Multiplicative noise (scale)
                     for (int i = 0; i < 4; i++) {
-                        float scale = ((float)arc4random() / UINT32_MAX) * 2.0f; // Scale range [0, 2]
+                        float scale = (float)(((double)arc4random() / UINT32_MAX) * 2.0); // Scale range [0, 2]
                         rawData[pixelIndex + i] *= scale;
                     }
                     break;
